+++ src/objects/index.js
@@ -12,3 +12,4 @@ export {default as WikiPage} from './WikiPage.js';
 export {default as Listing} from './Listing.js';
 export {default as More} from './More.js';
 export {default as UserList} from './UserList.js';
+export {default as ModmailConversation} from './ModmailConversation.js';